backend/usermanagement/requirements.txt
asgiref
Django
sqlparse
typing-extensions
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-allauth
requests
cryptography
